# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Serwish Api",
        "description": "Serwish description",
        "contact": {
            "email": "dev@serwish.ge"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://serwish.ge",
            "description": "serwish web-site"
        },
        {
            "url": "https://api.serwish.ge/api",
            "description": "serwish api"
        }
    ],
    "paths": {
...
        "/api/categories/list": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get categories by type",
                "operationId": "getFilteredCategories",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/FilterableSupportRequest"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiRequestException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
            "FilterableSupportRequest": {
                "properties": {
                    "locale": {
                        "description": "locale",
                        "format": "locale",
                        "default": "ka"
                    },
                    "type": {
                        "description": "type",
                        "format": "string",
                        "default": "POST"
                    },
                    "page": {
                        "description": "page",
                        "format": "integer",
                        "default": 1
                    },
                    "perPage": {
                        "description": "perPage",
                        "format": "integer",
                        "default": 20
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64"
                    },
                    "isActive": {
                        "title": "isActive",
                        "description": "isActive",
                        "format": "boolean"
                    },
                    "author_id": {
                        "title": "author_id",
                        "description": "author_id",
                        "format": "int64"
                    },
                    "type": {
                        "title": "type",
                        "description": "type",
                        "format": "string"
                    },
                    "translated": {
                        "$ref": "#/components/schemas/CategoryTranslation"
                    }
                },
                "type": "object"
            },
            "CategoryTranslation": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "format": "string"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "format": "string"
                    },
                    "slug": {
                        "title": "slug",
                        "description": "slug",
                        "format": "string"
                    },
                    "meta": {
                        "title": "meta",
                        "description": "meta",
                        "format": "string"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global Authorization: Bearer token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/token",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": {}
                    }
                },
                "scheme": "https"
            }
        }
    },
    "security": [
        []
    ]
}